begin improving logging and refactoring of tests
