+++ clients/python/test/e2e/ci/postprocess_e2e.py
@@ -1,20 +1,12 @@
-import shutil
 import warnings
 from pathlib import Path
 from typing import Set
 
 import pandas as pd
 import pytest
-import toml
 import typer
-from _utils import (
-    _ARTIFACTS_DIR,
-    _PYPROJECT_TOML,
-    ClientConfig,
-    E2eExitCodes,
-    E2eScriptFailure,
-    ServerConfig,
-)
+from _data_classes import Artifacts, ClientConfig, PytestIniFile, ServerConfig
+from _utils import E2eExitCodes, E2eScriptFailure
 from pydantic import ValidationError
 
 
@@ -47,22 +39,18 @@ def main(exit_code: int) -> None:
         )
         typer.Exit(code=E2eExitCodes.CI_SCRIPT_FAILURE)
 
-    _ARTIFACTS_DIR.mkdir(parents=True, exist_ok=True)
-    if not _PYPROJECT_TOML.is_file():
-        warnings.warn(f"cfg_file={_PYPROJECT_TOML}", E2eScriptFailure)
-        raise typer.Exit(code=E2eExitCodes.CI_SCRIPT_FAILURE)
-
-    # extract values
+    # get config
     try:
-        server_cfg: ServerConfig = ServerConfig(**toml.load(_PYPROJECT_TOML)["server"])
-        client_cfg: ClientConfig = ClientConfig(**toml.load(_PYPROJECT_TOML)["client"])
+        pytest_ini: PytestIniFile = PytestIniFile.read()
     except (ValueError, ValidationError):
-        print(toml.load(_PYPROJECT_TOML)["server"])
-        print(print(toml.load(_PYPROJECT_TOML)["client"]))
         raise typer.Exit(code=E2eExitCodes.INVALID_JSON_DATA)
 
+    client_cfg: ClientConfig = pytest_ini.client
+    server_cfg: ServerConfig = pytest_ini.server
+    artifacts: Artifacts = pytest_ini.artifacts
+
     # add result to json
-    result_file: Path = _ARTIFACTS_DIR / (client_cfg.client_ref + ".json")
+    result_file: Path = artifacts.result_data_frame
     new_df: pd.DataFrame = pd.DataFrame(
         columns=[client_cfg.client_ref], index=[server_cfg.url.netloc], data=[exit_code]
     )
@@ -75,12 +63,9 @@ def main(exit_code: int) -> None:
         result_df = new_df
     result_file.write_text(result_df.to_json())
 
-    # copy toml to artifacts dir
-    toml_dir: Path = _ARTIFACTS_DIR / (
-        client_cfg.client_ref + "+" + server_cfg.url.netloc
-    )
-    toml_dir.mkdir(exist_ok=False)
-    shutil.move(_PYPROJECT_TOML, toml_dir / _PYPROJECT_TOML.name)
+    # copy ini to artifacts dir
+    artifacts.log_dir.mkdir(exist_ok=True)
+    pytest_ini.generate(artifacts.log_dir / "pytest.ini")
     raise typer.Exit(code=pytest.ExitCode.OK)
